+++ noodles-sam/src/header/record/value.rs
@@ -1,5 +1,9 @@
 //! SAM header record value.
 
+pub mod map;
+
+pub use self::map::Map;
+
 use std::{error, fmt};
 
 use indexmap::IndexMap;